Refactor transmission model (#298)
* remove ununsed files

* merge EmergenceModel cpp into header

* merge MosqTrqnsmission + fitting into AnophelesModel

* remove MosqTransmission.h

* remove SimpleMPDEmergence.cpp  (merged into header)
