+++ model/CMakeLists.txt
@@ -46,10 +46,10 @@ set (Model_CPP
   Transmission/VectorModel.cpp
   Transmission/PerHost.cpp
   Transmission/Anopheles/AnophelesModel.cpp
-  Transmission/Anopheles/EmergenceModel.cpp
-  Transmission/Anopheles/MosqTransmission.cpp
-  Transmission/Anopheles/FixedEmergence.cpp
-  Transmission/Anopheles/SimpleMPDEmergence.cpp
+  #Transmission/Anopheles/EmergenceModel.cpp
+  #Transmission/Anopheles/MosqTransmission.cpp
+  #Transmission/Anopheles/FixedEmergence.cpp
+  #Transmission/Anopheles/SimpleMPDEmergence.cpp
   #Transmission/Anopheles/LCEmergence.cpp
   #Transmission/Anopheles/LifeCycle.cpp
   #Transmission/Anopheles/ResourceFitter.cpp
