+++ model/Transmission/Anopheles/EmergenceModel.h
@@ -27,13 +27,18 @@
 #include "util/checkpoint_containers.h"
 #include "util/vecDay.h"
 
+#include <cmath>
+
+#include "util/vectors.h"
+#include "util/CommandLine.h"
+#include "util/errors.h"
+
 namespace OM {
 namespace Transmission {
 namespace Anopheles {
 
 using namespace std;
-using util::vecDay;
-using util::LocalRng;
+using namespace OM::util;
 
 // forward declare to avoid circular dependency:
 class MosqTransmission;
@@ -49,34 +54,8 @@ class EmergenceModel
 public:
     ///@brief Initialisation and destruction
     //@{
-    EmergenceModel();
+    EmergenceModel() {}
     virtual ~EmergenceModel() {}
-    
-    /** Called to initialise life-cycle parameters from XML data.
-     * 
-     * Default implementation throws an assertion: only call if the life cycle
-     * model is used. */
-    virtual void initLifeCycle(const scnXml::LifeCycle& lcData){
-        assert (false);
-    }
-    
-    /** Called by initialise function to init variables directly related to EIR
-     * 
-     * @param anoph Data from XML
-     * @param initialisationEIR In/out parameter: TransmissionModel::initialisationEIR
-     * @param EIPDuration parameter from MosqTransmission (used for an estimation)
-     */
-    void initEIR(
-        const scnXml::AnophelesParams& anoph,
-        vector<double>& initialisationEIR,
-        SimTime EIPDuration );
-    
-    /** Set up the non-host-specific interventions. */
-    void initVectorInterv( const scnXml::VectorSpeciesIntervention& elt, size_t instance );
-    
-    /** Scale the internal EIR representation by factor; used as part of
-     * initialisation. */
-    void scaleEIR( double factor );
     
     /** Latter part of AnophelesModel::init2.
      *
@@ -86,18 +65,15 @@ public:
      * @param EIRtoS_v multiplication factor to convert input EIR into required
      * @param transmission reference to MosqTransmission object
      * S_v. */
-    virtual void init2( double tsP_A, double tsP_Amu, double tsP_A1, double tsP_Ah, double tsP_df, double tsP_dff, double EIRtoS_v, MosqTransmission& transmission ) =0;
+    virtual void init2(double tsP_dff, double initNv0FromSv, const vecDay<double>& forcedS_v, const vecDay<double>& mosqEmergeRate, const SimTime &mosqRestDuration) =0;
     
     /** Work out whether another interation is needed for initialisation and if
      * so, make necessary changes.
      *
      * @returns true if another iteration is needed. */
-    virtual bool initIterate (MosqTransmission& transmission) =0;
+    virtual void initIterate (double factor, const vecDay<double>& mosqEmergeRate) = 0;
     //@}
-    
-    /// Update per time-step (for larviciding intervention). Call before
-    /// getting emergence each time-step.
-    void update ();
+
     
     /** Model updates.
      * 
@@ -112,94 +88,34 @@ public:
      * @returns The number of adults emerging between the last simulated time
      * point and the one being calculated.
      */
-    virtual double update( SimTime d0, double nOvipositing, double S_v ) =0;
-    
-    ///@brief Interventions and reporting
-    //@{
-    /// Start an intervention affecting the vector population.
-    void deployVectorPopInterv (LocalRng& rng, size_t instance);
-    
+    virtual double update(const SimTime &d0, const vecDay<double>& mosqEmergeRate, double nOvipositing) =0;
+
     virtual double getResAvailability() const =0;
     virtual double getResRequirements() const =0;
-    //@}
-    
-    /// Checkpointing
+
     template<class S>
-    void operator& (S& stream) {
-        EIRRotateAngle & stream;
-        FSRotateAngle & stream;
-        FSCoeffic & stream;
-        forcedS_v & stream;
-        initNvFromSv & stream;
-        initOvFromSv & stream;
-        emergenceReduction & stream;
-        emergenceSurvival & stream;
-        checkpoint (stream);
-    }
+    void operator& (S& stream) { checkpoint (stream); }
     
 protected:
-    /** Return the proportion of emerging larvae to survive intervention
-     * effects. Should be between 0 and 1. */
-    inline double interventionSurvival() const{ return emergenceSurvival; }
+    // /** Return the proportion of emerging larvae to survive intervention
+    //  * effects. Should be between 0 and 1. */
+    // inline double interventionSurvival() const{ return emergenceSurvival; }
     
     virtual void checkpoint (istream& stream) =0;
     virtual void checkpoint (ostream& stream) =0;
-    
-    
-    // -----  parameters (constant after initialisation)  -----
-    
-    ///@brief Descriptions of transmission, used primarily during warmup
-    //@{
-    /// Angle (in radians) to rotate series generated by FSCoeffic by, for EIR.
-    double EIRRotateAngle;
 
-    /// Rotation angle (in radians) for emergence rate. Both offset for EIR given in XML file and
-    /// offset needed to fit target EIR (delayed from emergence rate). Checkpoint.
-    double FSRotateAngle;
-
-    /** Fourier coefficients for EIR / forcedS_v series, input from XML file.
-     *
-     * Initially used to calculate initialisation EIR, then scaled to calc. S_v.
-     *
-     * When calcFourierEIR is used to produce an EIR from this over 365
-     * (365) elements, the resulting EIR has units of
-     * infectious bites per adult per day.
-     *
-     * fcEir must have odd length and is ordered: [a0, a1, b1, ..., an, bn].
-     * FSCoeffic[0] needs checkpointing, the rest doesn't. */
-    vector<double> FSCoeffic;
-
-    /** S_v used to force an EIR during vector init.
+private:
+    /** Emergence rate of new mosquitoes, for every day of the year (N_v0).
      * 
      * Has annual periodicity: length is 365. First value (index 0) corresponds
      * to first day of year (1st Jan or something else if rebased). In 5-day
      * time-step model values at indecies 0 through 4 are used to calculate the
      * state at time-step 1.
+     * 
+     * Units: Animals per day.
      *
      * Should be checkpointed. */
-    vecDay<double> forcedS_v;
-    
-    /** Conversion factor from forcedS_v to (initial values of) N_v (1 / ρ_S).
-     * Should be checkpointed. */
-    double initNvFromSv;
-    
-    /** Conversion factor from forcedS_v to (initial values of) O_v (ρ_O / ρ_S).
-     * Should be checkpointed. */
-    double initOvFromSv;
-    //@}
-    
-    /** @brief Intervention parameters
-     *
-     * Checkpointed. */
-    //@{
-    /// Description of intervention killing effects on emerging pupae
-    vector<util::SimpleDecayingValue> emergenceReduction;
-    /// Cache parameter updated by update()
-    double emergenceSurvival;   // survival with regards to intervention effects
-    //@}
-
-    double scaleFactor, shiftAngle;
-    bool rotated, scaled;
+    vecDay<double> mosqEmergeRate;
 };
 
 }
