+++ test/expected/outputLifeNet2.txt
@@ -1,44 +1,44 @@
 1	1	0	500
-1	1	3	2
-1	0	7	0.000558112
-1	1	11	209
-1	1	12	19
-1	1	13	3
-1	1	14	584
-1	1	15	4
+1	1	3	6
+1	0	7	0.00180679
+1	1	11	1084
+1	1	12	101
+1	1	13	13
+1	1	14	3012
+1	1	15	24
 1	1	16	0
-1	1	17	1
-1	1	18	3
-1	1	19	1
-1	1	23	2
+1	1	17	2
+1	1	18	2
+1	1	19	4
+1	1	23	11
 1	1	24	0
-1	0	26	1.52583e-05
+1	0	26	0.00534679
 1	0	35	0.0136986
-1	0	36	0.00226842
+1	0	36	0.0136047
 2	1	0	500
-2	1	3	0
-2	0	7	0
-2	1	11	3
-2	1	12	0
+2	1	3	3
+2	0	7	0.000283843
+2	1	11	25
+2	1	12	8
 2	1	13	0
-2	1	14	9
-2	1	15	0
+2	1	14	70
+2	1	15	1
 2	1	16	0
 2	1	17	0
-2	1	18	0
+2	1	18	1
 2	1	19	0
 2	1	23	0
 2	1	24	0
-2	0	26	8.66873e-05
+2	0	26	0.00301962
 2	0	35	0.0136986
-2	0	36	0.000126053
+2	0	36	0.00242098
 3	1	0	500
 3	1	3	0
 3	0	7	0
-3	1	11	0
+3	1	11	1
 3	1	12	0
 3	1	13	0
-3	1	14	0
+3	1	14	7
 3	1	15	0
 3	1	16	0
 3	1	17	0
@@ -46,16 +46,16 @@
 3	1	19	0
 3	1	23	0
 3	1	24	0
-3	0	26	1.34916e-13
+3	0	26	1.27998e-05
 3	0	35	0.0136986
-3	0	36	0
+3	0	36	3.71572e-05
 4	1	0	500
 4	1	3	0
 4	0	7	0
 4	1	11	0
 4	1	12	0
 4	1	13	0
-4	1	14	0
+4	1	14	2
 4	1	15	0
 4	1	16	0
 4	1	17	0
@@ -63,16 +63,16 @@
 4	1	19	0
 4	1	23	0
 4	1	24	0
-4	0	26	0
+4	0	26	1.52512e-05
 4	0	35	0.0136986
-4	0	36	0
+4	0	36	2.77667e-05
 5	1	0	500
 5	1	3	0
-5	0	7	9.52132e-06
-5	1	11	1
-5	1	12	1
+5	0	7	0
+5	1	11	0
+5	1	12	0
 5	1	13	0
-5	1	14	8
+5	1	14	0
 5	1	15	0
 5	1	16	0
 5	1	17	0
@@ -80,12 +80,12 @@
 5	1	19	0
 5	1	23	0
 5	1	24	0
-5	0	26	0.000391624
+5	0	26	0
 5	0	35	0.0136986
-5	0	36	0.000170491
+5	0	36	0
 6	1	0	500
 6	1	3	0
-6	0	7	1.69827e-06
+6	0	7	0
 6	1	11	0
 6	1	12	0
 6	1	13	0
@@ -97,16 +97,16 @@
 6	1	19	0
 6	1	23	0
 6	1	24	0
-6	0	26	2.46337e-06
+6	0	26	0
 6	0	35	0.0136986
-6	0	36	2.14909e-06
+6	0	36	0
 7	1	0	500
 7	1	3	0
 7	0	7	0
 7	1	11	0
 7	1	12	0
 7	1	13	0
-7	1	14	0
+7	1	14	1
 7	1	15	0
 7	1	16	0
 7	1	17	0
@@ -114,16 +114,16 @@
 7	1	19	0
 7	1	23	0
 7	1	24	0
-7	0	26	2.90818e-08
+7	0	26	1.68273e-05
 7	0	35	0.0136986
-7	0	36	0
+7	0	36	7.37254e-06
 8	1	0	500
-8	1	3	0
-8	0	7	0
-8	1	11	0
+8	1	3	3
+8	0	7	0.00190743
+8	1	11	5
 8	1	12	0
 8	1	13	0
-8	1	14	0
+8	1	14	15
 8	1	15	0
 8	1	16	0
 8	1	17	0
@@ -131,16 +131,16 @@
 8	1	19	0
 8	1	23	0
 8	1	24	0
-8	0	26	0
+8	0	26	0.00075738
 8	0	35	0.0136986
-8	0	36	0
+8	0	36	0.000305984
 9	1	0	500
 9	1	3	0
 9	0	7	0
-9	1	11	0
-9	1	12	0
+9	1	11	5
+9	1	12	2
 9	1	13	0
-9	1	14	4
+9	1	14	10
 9	1	15	0
 9	1	16	0
 9	1	17	0
@@ -148,9 +148,9 @@
 9	1	19	0
 9	1	23	0
 9	1	24	0
-9	0	26	3.8428e-08
+9	0	26	0.000439465
 9	0	35	0.0136986
-9	0	36	0
+9	0	36	0.000288537
 10	1	0	500
 10	1	3	0
 10	0	7	0
@@ -171,10 +171,10 @@
 11	1	0	500
 11	1	3	0
 11	0	7	0
-11	1	11	1
+11	1	11	0
 11	1	12	0
 11	1	13	0
-11	1	14	2
+11	1	14	0
 11	1	15	0
 11	1	16	0
 11	1	17	0
@@ -182,9 +182,9 @@
 11	1	19	0
 11	1	23	0
 11	1	24	0
-11	0	26	9.65227e-06
+11	0	26	0
 11	0	35	0.0136986
-11	0	36	1.49426e-05
+11	0	36	0
 12	1	0	500
 12	1	3	0
 12	0	7	0
@@ -205,10 +205,10 @@
 13	1	0	500
 13	1	3	0
 13	0	7	0
-13	1	11	1
+13	1	11	0
 13	1	12	0
 13	1	13	0
-13	1	14	2
+13	1	14	0
 13	1	15	0
 13	1	16	0
 13	1	17	0
@@ -216,16 +216,16 @@
 13	1	19	0
 13	1	23	0
 13	1	24	0
-13	0	26	0.000111331
+13	0	26	0
 13	0	35	0.0136986
-13	0	36	2.18898e-05
+13	0	36	0
 14	1	0	500
 14	1	3	0
 14	0	7	0
 14	1	11	0
 14	1	12	0
 14	1	13	0
-14	1	14	2
+14	1	14	0
 14	1	15	0
 14	1	16	0
 14	1	17	0
@@ -233,16 +233,16 @@
 14	1	19	0
 14	1	23	0
 14	1	24	0
-14	0	26	9.79172e-06
+14	0	26	0
 14	0	35	0.0136986
-14	0	36	1.88809e-05
+14	0	36	0
 15	1	0	500
 15	1	3	0
 15	0	7	0
-15	1	11	3
+15	1	11	0
 15	1	12	0
 15	1	13	0
-15	1	14	4
+15	1	14	0
 15	1	15	0
 15	1	16	0
 15	1	17	0
@@ -250,16 +250,16 @@
 15	1	19	0
 15	1	23	0
 15	1	24	0
-15	0	26	7.42558e-05
+15	0	26	0
 15	0	35	0.0136986
-15	0	36	9.8369e-07
+15	0	36	0
 16	1	0	500
-16	1	3	1
-16	0	7	0.000355984
-16	1	11	2
+16	1	3	0
+16	0	7	0
+16	1	11	0
 16	1	12	0
 16	1	13	0
-16	1	14	5
+16	1	14	0
 16	1	15	0
 16	1	16	0
 16	1	17	0
@@ -267,16 +267,16 @@
 16	1	19	0
 16	1	23	0
 16	1	24	0
-16	0	26	1.12935e-05
+16	0	26	0
 16	0	35	0.0136986
-16	0	36	8.43746e-07
+16	0	36	0
 17	1	0	500
 17	1	3	0
 17	0	7	0
-17	1	11	1
+17	1	11	0
 17	1	12	0
 17	1	13	0
-17	1	14	2
+17	1	14	1
 17	1	15	0
 17	1	16	0
 17	1	17	0
@@ -284,16 +284,16 @@
 17	1	19	0
 17	1	23	0
 17	1	24	0
-17	0	26	1.89583e-05
+17	0	26	4.00806e-07
 17	0	35	0.0136986
-17	0	36	2.1594e-05
+17	0	36	8.33187e-07
 18	1	0	500
 18	1	3	0
-18	0	7	4.34977e-07
+18	0	7	0
 18	1	11	0
 18	1	12	0
 18	1	13	0
-18	1	14	1
+18	1	14	0
 18	1	15	0
 18	1	16	0
 18	1	17	0
@@ -301,16 +301,16 @@
 18	1	19	0
 18	1	23	0
 18	1	24	0
-18	0	26	1.41966e-07
+18	0	26	0
 18	0	35	0.0136986
 18	0	36	0
 19	1	0	500
 19	1	3	0
 19	0	7	0
-19	1	11	2
-19	1	12	1
+19	1	11	1
+19	1	12	0
 19	1	13	0
-19	1	14	2
+19	1	14	3
 19	1	15	0
 19	1	16	0
 19	1	17	0
@@ -318,16 +318,16 @@
 19	1	19	0
 19	1	23	0
 19	1	24	0
-19	0	26	0.00038282
+19	0	26	2.52103e-05
 19	0	35	0.0136986
-19	0	36	6.47918e-05
+19	0	36	4.17095e-05
 20	1	0	500
 20	1	3	0
-20	0	7	0
-20	1	11	0
+20	0	7	2.50993e-05
+20	1	11	1
 20	1	12	0
 20	1	13	0
-20	1	14	0
+20	1	14	1
 20	1	15	0
 20	1	16	0
 20	1	17	0
@@ -335,16 +335,16 @@
 20	1	19	0
 20	1	23	0
 20	1	24	0
-20	0	26	0
+20	0	26	9.86987e-07
 20	0	35	0.0136986
-20	0	36	0
+20	0	36	3.61435e-07
 21	1	0	500
 21	1	3	0
 21	0	7	0
-21	1	11	0
+21	1	11	1
 21	1	12	0
 21	1	13	0
-21	1	14	0
+21	1	14	1
 21	1	15	0
 21	1	16	0
 21	1	17	0
@@ -352,9 +352,9 @@
 21	1	19	0
 21	1	23	0
 21	1	24	0
-21	0	26	0
+21	0	26	2.94491e-06
 21	0	35	0.0136986
-21	0	36	0
+21	0	36	4.0736e-07
 22	1	0	500
 22	1	3	0
 22	0	7	0
@@ -395,7 +395,7 @@
 24	1	11	0
 24	1	12	0
 24	1	13	0
-24	1	14	0
+24	1	14	1
 24	1	15	0
 24	1	16	0
 24	1	17	0
@@ -403,9 +403,9 @@
 24	1	19	0
 24	1	23	0
 24	1	24	0
-24	0	26	0
+24	0	26	9.63045e-06
 24	0	35	0.0136986
-24	0	36	0
+24	0	36	3.25113e-07
 25	1	0	500
 25	1	3	0
 25	0	7	0
