+++ test/expected/outputSimpleMPDTest.txt
@@ -1,58 +1,58 @@
-1	1	0	4
-1	2	0	5
-1	3	0	11
-1	4	0	12
-1	5	0	22
-1	6	0	16
-1	7	0	15
-1	8	0	5
-1	9	0	6
-1	10	0	3
-1	11	0	1
-1	1	2	3.49607
-1	2	2	5
-1	3	2	11
-1	4	2	12
-1	5	2	22
-1	6	2	16
-1	7	2	15
-1	8	2	5
-1	9	2	6
-1	10	2	3
-1	11	2	1
+1	1	0	5
+1	2	0	3
+1	3	0	10
+1	4	0	16
+1	5	0	19
+1	6	0	18
+1	7	0	10
+1	8	0	12
+1	9	0	3
+1	10	0	2
+1	11	0	2
+1	1	2	3.79622
+1	2	2	3
+1	3	2	10
+1	4	2	16
+1	5	2	19
+1	6	2	18
+1	7	2	10
+1	8	2	12
+1	9	2	3
+1	10	2	2
+1	11	2	2
 1	1	3	3
-1	2	3	5
-1	3	3	10
-1	4	3	12
-1	5	3	21
-1	6	3	15
-1	7	3	13
-1	8	3	5
-1	9	3	6
-1	10	3	3
-1	11	3	1
-1	1	4	27.112
-1	2	4	38.6666
-1	3	4	83.0301
-1	4	4	93.0087
-1	5	4	168.584
-1	6	4	122.707
-1	7	4	106.134
-1	8	4	38.4537
-1	9	4	45.6642
-1	10	4	23.2421
-1	11	4	7.75461
-1	1	5	15.7884
-1	2	5	35.6944
-1	3	5	45.5053
-1	4	5	61.3741
-1	5	5	72.4117
-1	6	5	62.0851
-1	7	5	47.5588
-1	8	5	13.5111
-1	9	5	17.5212
-1	10	5	13.8329
-1	11	5	4.94061
+1	2	3	3
+1	3	3	9
+1	4	3	16
+1	5	3	19
+1	6	3	16
+1	7	3	9
+1	8	3	12
+1	9	3	3
+1	10	3	2
+1	11	3	2
+1	1	4	33.0457
+1	2	4	22.7222
+1	3	4	76.5771
+1	4	4	123.316
+1	5	4	146.651
+1	6	4	133.137
+1	7	4	72.7192
+1	8	4	91.7209
+1	9	4	23.263
+1	10	4	15.4842
+1	11	4	15.5073
+1	1	5	20.6904
+1	2	5	23.401
+1	3	5	51.9769
+1	4	5	75.2709
+1	5	5	77.613
+1	6	5	71.643
+1	7	5	20.4968
+1	8	5	41.1681
+1	9	5	16.8845
+1	10	5	7.9303
+1	11	5	9.47007
 1	1	11	0
 1	2	11	0
 1	3	11	0
@@ -86,85 +86,85 @@
 1	9	13	0
 1	10	13	0
 1	11	13	0
-1	1	30	1.13146
-1	2	30	1.7254
-1	3	30	4.93568
-1	4	30	7.36517
-1	5	30	20.846
-1	6	30	19.5223
-1	7	30	18.5466
-1	8	30	6.14927
-1	9	30	7.40851
-1	10	30	3.56869
-1	11	30	1.36234
-1	1	31	163.818
-1	1	32	1765
-1	1	33	185.877
-1	1	34	101.253
+1	1	30	1.65471
+1	2	30	1.21315
+1	3	30	5.601
+1	4	30	12.0771
+1	5	30	22.4577
+1	6	30	26.514
+1	7	30	14.8742
+1	8	30	17.7873
+1	9	30	4.28638
+1	10	30	2.92351
+1	11	30	3.08225
+1	1	31	179.301
+1	1	32	1924.55
+1	1	33	226.48
+1	1	34	122.215
 1	0	35	1.44137
-1	0	36	1.22952
+1	0	36	1.47803
 2	1	0	8
 2	2	0	2
-2	3	0	6
-2	4	0	15
-2	5	0	26
-2	6	0	10
-2	7	0	17
-2	8	0	10
-2	9	0	3
-2	10	0	2
+2	3	0	10
+2	4	0	13
+2	5	0	18
+2	6	0	21
+2	7	0	15
+2	8	0	3
+2	9	0	5
+2	10	0	4
 2	11	0	1
-2	1	2	6.35629
+2	1	2	7.19586
 2	2	2	2
-2	3	2	6
-2	4	2	15
-2	5	2	26
-2	6	2	10
-2	7	2	17
-2	8	2	10
-2	9	2	3
-2	10	2	2
+2	3	2	10
+2	4	2	13
+2	5	2	18
+2	6	2	21
+2	7	2	15
+2	8	2	3
+2	9	2	5
+2	10	2	4
 2	11	2	1
 2	1	3	6
 2	2	3	2
-2	3	3	6
-2	4	3	13
-2	5	3	25
-2	6	3	10
-2	7	3	16
-2	8	3	9
-2	9	3	2
-2	10	3	2
+2	3	3	10
+2	4	3	12
+2	5	3	18
+2	6	3	21
+2	7	3	13
+2	8	3	3
+2	9	3	5
+2	10	3	4
 2	11	3	1
-2	1	4	55.5681
-2	2	4	15.4832
-2	3	4	46.4968
-2	4	4	111.681
-2	5	4	199.978
-2	6	4	76.6654
-2	7	4	125.187
-2	8	4	72.564
-2	9	4	22.5673
-2	10	4	15.5144
-2	11	4	7.75493
-2	1	5	31.8279
-2	2	5	8.65613
-2	3	5	40.9986
-2	4	5	55.6484
-2	5	5	109.123
-2	6	5	29.8537
-2	7	5	65.0375
-2	8	5	19.5937
-2	9	5	7.99836
-2	10	5	12.8465
-2	11	5	5.02464
-2	1	11	1
-2	2	11	8
-2	3	11	7
-2	4	11	3
+2	1	4	54.6646
+2	2	4	15.4271
+2	3	4	77.1586
+2	4	4	99.9746
+2	5	4	138.523
+2	6	4	161.638
+2	7	4	105.757
+2	8	4	22.7844
+2	9	4	37.9751
+2	10	4	30.9925
+2	11	4	7.75463
+2	1	5	28.5093
+2	2	5	16.3609
+2	3	5	43.5042
+2	4	5	60.1492
+2	5	5	58.5319
+2	6	5	98.903
+2	7	5	34.0672
+2	8	5	4.65075
+2	9	5	15.677
+2	10	5	17.7355
+2	11	5	5.04781
+2	1	11	3
+2	2	11	1
+2	3	11	2
+2	4	11	1
 2	5	11	0
-2	6	11	0
-2	7	11	2
+2	6	11	1
+2	7	11	1
 2	8	11	0
 2	9	11	0
 2	10	11	0
@@ -180,7 +180,7 @@
 2	9	12	0
 2	10	12	0
 2	11	12	0
-2	1	13	2
+2	1	13	0
 2	2	13	0
 2	3	13	0
 2	4	13	0
@@ -191,84 +191,84 @@
 2	9	13	0
 2	10	13	0
 2	11	13	0
-2	1	30	971.154
-2	2	30	670.911
-2	3	30	2502.32
-2	4	30	4907.69
-2	5	30	11859.6
-2	6	30	7898.11
-2	7	30	9865.62
-2	8	30	5169.08
-2	9	30	2972.58
-2	10	30	1049.38
-2	11	30	842.416
-2	1	31	163.823
-2	1	32	1763.35
-2	1	33	194.799
-2	1	34	106.085
+2	1	30	1219.79
+2	2	30	747.047
+2	3	30	2255.48
+2	4	30	5560.91
+2	5	30	12784.2
+2	6	30	14185.1
+2	7	30	10401.3
+2	8	30	6494.51
+2	9	30	3289.72
+2	10	30	1877.55
+2	11	30	1084.46
+2	1	31	179.289
+2	1	32	1927.81
+2	1	33	208.526
+2	1	34	113.684
 2	0	35	2.44665
-2	0	36	2.25118
-3	1	0	7
-3	2	0	7
-3	3	0	5
-3	4	0	13
-3	5	0	24
-3	6	0	13
-3	7	0	12
-3	8	0	13
-3	9	0	3
-3	10	0	2
+2	0	36	2.68917
+3	1	0	6
+3	2	0	4
+3	3	0	9
+3	4	0	11
+3	5	0	21
+3	6	0	20
+3	7	0	14
+3	8	0	4
+3	9	0	6
+3	10	0	4
 3	11	0	1
-3	1	2	5.50045
-3	2	2	7
-3	3	2	5
-3	4	2	13
-3	5	2	24
-3	6	2	13
-3	7	2	12
-3	8	2	13
-3	9	2	3
-3	10	2	2
+3	1	2	5.12148
+3	2	2	4
+3	3	2	9
+3	4	2	11
+3	5	2	21
+3	6	2	20
+3	7	2	14
+3	8	2	4
+3	9	2	6
+3	10	2	4
 3	11	2	1
 3	1	3	5
-3	2	3	6
-3	3	3	5
-3	4	3	12
-3	5	3	23
-3	6	3	13
-3	7	3	11
-3	8	3	12
-3	9	3	3
-3	10	3	2
+3	2	3	3
+3	3	3	9
+3	4	3	11
+3	5	3	20
+3	6	3	19
+3	7	3	12
+3	8	3	4
+3	9	3	5
+3	10	3	4
 3	11	3	1
-3	1	4	46.8425
-3	2	4	53.2117
-3	3	4	38.75
-3	4	4	98.9144
-3	5	4	184.459
-3	6	4	99.7813
-3	7	4	86.7993
-3	8	4	95.4213
-3	9	4	22.5699
-3	10	4	15.5144
-3	11	4	7.75501
-3	1	5	12.2955
-3	2	5	29.5776
-3	3	5	31.4108
-3	4	5	52.6896
-3	5	5	100.118
-3	6	5	34.8049
-3	7	5	55.9692
-3	8	5	27.252
-3	9	5	7.27032
-3	10	5	12.6396
-3	11	5	4.97909
-3	1	11	3
+3	1	4	41.1997
+3	2	4	30.0615
+3	3	4	68.946
+3	4	4	85.2193
+3	5	4	161.019
+3	6	4	153.748
+3	7	4	98.2055
+3	8	4	30.305
+3	9	4	45.6888
+3	10	4	30.9903
+3	11	4	7.75455
+3	1	5	29.2331
+3	2	5	16.5688
+3	3	5	39.4396
+3	4	5	57.7202
+3	5	5	71.0403
+3	6	5	89.6877
+3	7	5	34.2291
+3	8	5	7.12423
+3	9	5	18.3781
+3	10	5	17.3406
+3	11	5	4.90078
+3	1	11	1
 3	2	11	0
-3	3	11	1
+3	3	11	0
 3	4	11	0
 3	5	11	0
-3	6	11	0
+3	6	11	1
 3	7	11	1
 3	8	11	0
 3	9	11	0
@@ -285,8 +285,8 @@
 3	9	12	0
 3	10	12	0
 3	11	12	0
-3	1	13	1
-3	2	13	2
+3	1	13	0
+3	2	13	1
 3	3	13	0
 3	4	13	0
 3	5	13	0
@@ -296,20 +296,20 @@
 3	9	13	0
 3	10	13	0
 3	11	13	0
-3	1	30	249.945
-3	2	30	193.242
-3	3	30	348.774
-3	4	30	1165.25
-3	5	30	3105.64
-3	6	30	1932.4
-3	7	30	2346.22
-3	8	30	1806.4
-3	9	30	482.658
-3	10	30	307.683
-3	11	30	152.242
-3	1	31	163.825
-3	1	32	1763.03
-3	1	33	187.465
-3	1	34	102.429
+3	1	30	265.204
+3	2	30	189.498
+3	3	30	608.525
+3	4	30	1010.35
+3	5	30	2792.52
+3	6	30	3663.28
+3	7	30	2801.96
+3	8	30	617.747
+3	9	30	907.627
+3	10	30	698.639
+3	11	30	199.96
+3	1	31	179.309
+3	1	32	1925.88
+3	1	33	203.139
+3	1	34	111.317
 3	0	35	2.44665
-3	0	36	2.1823
+3	0	36	2.47293
